requirements.txt
coverage==4.5.1
Django==2.1.2
django-nose==1.4.6
djangorestframework==3.9.0
nose==1.3.7
pytz==2018.9
virtualenv==16.0.0
